src/anuncios/requirements.txt
django
mysqlclient
djangorestframework
django-filter
drf-yasg
coreapi
rest_framework_swagger